Refactored App into Map with changes
