+++ src/ui/components/Options.jsx
@@ -16,7 +16,7 @@ function Options(props) {
 }
 
 Options.propTypes = {
-  children: PropTypes.element.isRequired,
+  children: PropTypes.arrayOf(PropTypes.element).isRequired,
 };
 
 export default Options;
