+++ src/ui/utils/values.js
@@ -1,0 +1,3 @@
+export default {
+  drawerWidth: 240,
+};
